third_party/blink/web_tests/fast/forms/textarea/textarea-set-defaultvalue-after-value-expected.txt

Tests to check behaviors of dirty flag and defaultValue change.

PASS document.execCommand("InsertText", false, "user-updated"); textarea.value is "user-updated"
PASS textarea.defaultValue = "default 2"; textarea.value is "user-updated"
PASS form.reset(); textarea.value is "default 2"
PASS textarea.defaultValue = "default 3"; textarea.value is "default 3"
PASS textarea.value = "script-updated"; textarea.value is "script-updated"
PASS textarea.defaultValue = "default 4"; textarea.value is "script-updated"
PASS form.reset(); textarea.value is "default 4"
PASS textarea.value = "script-updated"; textarea.value is "script-updated"
PASS textarea.innerHTML = "default 5"; textarea.value is "script-updated"
PASS form.reset(); textarea.value is "default 5"
PASS textarea.innerHTML = "default 6"; textarea.value is "default 6"
PASS textarea.value = "script-updated"; textarea.value is "script-updated"
PASS textarea.removeChild(textarea.firstChild); textarea.appendChild(document.createTextNode("default 5")); textarea.value is "script-updated"
PASS form.reset(); textarea.value is "default 5"
PASS textarea.removeChild(textarea.firstChild); textarea.appendChild(document.createTextNode("default 6")); textarea.value is "default 6"
PASS successfullyParsed is true

TEST COMPLETE
